requirements.txt
requests>=2.28.0
python-dotenv>=0.21.0
PyGithub>=1.58.0
pydantic>=1.10.0
httpx>=0.23.0
aiohttp>=3.8.0
pytest>=7.0.0
pytest-cov>=4.0.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
